+++ src/Type/PaidMedia.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Type;
+
+/**
+ * @see https://core.telegram.org/bots/api#paidmedia
+ */
+interface PaidMedia
+{
+    public function getType(): string;
+}